backend/requirements.txt
flask==2.3.3 
flask-cors==4.0.0 
sqlalchemy==2.0.23 
pymysql==1.1.0 
mysql-connector-python==8.2.0 
requests==2.31.0 
python-dotenv==1.0.0 
